Check for definedness of php_register_mods_ini var
The defaults vars file sets the `php_modules_conf` var to an empty list,
but a role task assumes that the var will be defined, by using referencing
the skipped registered var that includes results captured via `with_items`.
The solution is to check for definedness and skip the dependent task if
the var isn't defined.